move lib folder to src.
