+++ __tests__/commit-file-async.spec.ts
@@ -4,7 +4,7 @@ import os from 'os';
 import path from 'path';
 import { create as createMemFs } from 'mem-fs';
 import sinon from 'sinon';
-import { type MemFsEditor, type MemFsEditorFile, create } from '../lib/index.js';
+import { type MemFsEditor, type MemFsEditorFile, create } from '../src/index.js';
 
 const rmSync = filesystem.rmSync || filesystem.rmdirSync;
