+++ __tests__/delete.spec.ts
@@ -1,6 +1,6 @@
 import { describe, beforeEach, it, expect } from 'vitest';
 import path from 'path';
-import { type MemFsEditor, create } from '../lib/index.js';
+import { type MemFsEditor, create } from '../src/index.js';
 import { create as createMemFs } from 'mem-fs';
 import { getFixture } from './fixtures.js';
