+++ __tests__/exists.spec.ts
@@ -1,5 +1,5 @@
 import { describe, beforeEach, it, expect } from 'vitest';
-import { type MemFsEditor, create } from '../lib/index.js';
+import { type MemFsEditor, create } from '../src/index.js';
 import { create as createMemFs } from 'mem-fs';
 import { getFixture } from './fixtures.js';
