+++ __tests__/move.spec.ts
@@ -1,7 +1,7 @@
 import { describe, beforeEach, it, expect } from 'vitest';
 import os from 'os';
 import path from 'path';
-import { type MemFsEditor, create } from '../lib/index.js';
+import { type MemFsEditor, create } from '../src/index.js';
 import { create as createMemFs } from 'mem-fs';
 import { getFixture } from './fixtures.js';
