+++ __tests__/read-json.spec.ts
@@ -1,6 +1,6 @@
 import { describe, beforeEach, it, expect } from 'vitest';
 import sinon from 'sinon';
-import { type MemFsEditor, create } from '../lib/index.js';
+import { type MemFsEditor, create } from '../src/index.js';
 import { create as createMemFs } from 'mem-fs';
 import escape from 'escape-regexp';
 import { getFixture } from './fixtures.js';
